+++ packages/react-docs/patternfly-docs.source.js
@@ -10,11 +10,19 @@ module.exports = (sourceMD, sourceProps) => {
   const reactTablePath = require.resolve('@patternfly/react-table/package.json').replace('package.json', 'src');
   const reactChartsPath = require.resolve('@patternfly/react-charts/package.json').replace('package.json', 'src');
   const reactDateTimePath = require.resolve('@patternfly/react-datetime/package.json').replace('package.json', 'src');
+  const reactCatalogViewPath = require
+    .resolve('@patternfly/react-catalog-view-extension/package.json')
+    .replace('package.json', 'src');
+  const reactVirtualizedPath = require
+    .resolve('@patternfly/react-virtualized-extension/package.json')
+    .replace('package.json', 'src');
   const reactPropsIgnore = '**/*.test.tsx';
   sourceProps(path.join(reactCorePath, '/**/*.tsx'), reactPropsIgnore);
   sourceProps(path.join(reactTablePath, '/**/*.tsx'), reactPropsIgnore);
   sourceProps(path.join(reactChartsPath, '/**/*.tsx'), reactPropsIgnore);
   sourceProps(path.join(reactDateTimePath, '/**/*.tsx'), reactPropsIgnore);
+  sourceProps(path.join(reactCatalogViewPath, '/**/*.tsx'), reactPropsIgnore);
+  sourceProps(path.join(reactVirtualizedPath, '/**/*.tsx'), reactPropsIgnore);
 
   // React MD
   sourceMD(path.join(reactCorePath, '/**/examples/*.md'), 'react');
@@ -29,6 +37,12 @@ module.exports = (sourceMD, sourceProps) => {
 
   // Date-time MD (no demos yet)
   sourceMD(path.join(reactDateTimePath, '/**/examples/*.md'), 'react');
+
+  // Catalog view MD
+  sourceMD(path.join(reactCatalogViewPath, '/**/examples/*.md'), 'react');
+
+  // Virtualized MD
+  sourceMD(path.join(reactVirtualizedPath, '/**/examples/*.md'), 'react');
 
   // Release notes
   sourceMD(require.resolve('@patternfly/react-docs/RELEASE-NOTES.md'), 'react');