+++ packages/react-table/src/components/Table/EditableSelectInputCell.tsx
@@ -4,7 +4,6 @@ import { Select, SelectOptionObject, SelectProps } from '@patternfly/react-core'
 import inlineStyles from '@patternfly/react-styles/css/components/InlineEdit/inline-edit';
 import formStyles from '@patternfly/react-styles/css/components/Form/form';
 
-// eslint-disable-next-line @typescript-eslint/interface-name-prefix
 export interface IEditableSelectInputCell extends Omit<React.HTMLProps<HTMLElement | HTMLDivElement>, 'onSelect'> {
   /** Row index of this select input cell */
   rowIndex: number;