+++ packages/react-catalog-view-extension/CHANGELOG.md
@@ -3,6 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.68](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.67...@patternfly/react-catalog-view-extension@4.8.68) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.67](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.66...@patternfly/react-catalog-view-extension@4.8.67) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.66](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.65...@patternfly/react-catalog-view-extension@4.8.66) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.65](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.64...@patternfly/react-catalog-view-extension@4.8.65) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.64](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.63...@patternfly/react-catalog-view-extension@4.8.64) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.63](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.62...@patternfly/react-catalog-view-extension@4.8.63) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.62](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.61...@patternfly/react-catalog-view-extension@4.8.62) (2020-10-07)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.61](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.60...@patternfly/react-catalog-view-extension@4.8.61) (2020-10-07)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.60](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.59...@patternfly/react-catalog-view-extension@4.8.60) (2020-10-06)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.59](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.58...@patternfly/react-catalog-view-extension@4.8.59) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.58](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.57...@patternfly/react-catalog-view-extension@4.8.58) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.57](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.56...@patternfly/react-catalog-view-extension@4.8.57) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
 ## [4.8.56](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.55...@patternfly/react-catalog-view-extension@4.8.56) (2020-10-02)
 
 **Note:** Version bump only for package @patternfly/react-catalog-view-extension