+++ packages/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.124](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.123...@patternfly/react-inline-edit-extension@4.5.124) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.123](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.122...@patternfly/react-inline-edit-extension@4.5.123) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.122](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.121...@patternfly/react-inline-edit-extension@4.5.122) (2020-10-12)
+
+
+### Bug Fixes
+
+* **inputgroup:** allow string children ([#4963](https://github.com/patternfly/patternfly-react/issues/4963)) ([2f90073](https://github.com/patternfly/patternfly-react/commit/2f90073111e01c4c5597d04b2a337707881a33ec))
+
+
+
+
+
+## [4.5.121](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.120...@patternfly/react-inline-edit-extension@4.5.121) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.120](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.119...@patternfly/react-inline-edit-extension@4.5.120) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.119](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.118...@patternfly/react-inline-edit-extension@4.5.119) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.118](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.117...@patternfly/react-inline-edit-extension@4.5.118) (2020-10-07)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.117](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.116...@patternfly/react-inline-edit-extension@4.5.117) (2020-10-06)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.116](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.115...@patternfly/react-inline-edit-extension@4.5.116) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.115](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.114...@patternfly/react-inline-edit-extension@4.5.115) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.114](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.113...@patternfly/react-inline-edit-extension@4.5.114) (2020-10-05)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [4.5.113](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.112...@patternfly/react-inline-edit-extension@4.5.113) (2020-10-02)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension